Cosas actualizadas pull necesario

+++ team_no_team_rocket/Puntuacion.java
@@ -40,7 +40,7 @@ public class Puntuacion {
 		this.puntosSobreCinco = puntosSobreCinco;
 	}
 
-	public int getVecesPuntuado() {
+	public double getVecesPuntuado() {
 		return vecesPuntuado;
 	}
